Moved charts to seperate component
